Modify demo to use Inmem Babble node
Until now, the demo was deploying evm-babble and babble in separate
processes to demonstrate the socket interface. We now have the
additional option to run evm-babble and babble in the same process,
by using the inmem interface. We modified the docker demo to use
evm-babble in inmmem mode. The next work is to allow choosing one
mode or the other from a configuration flag (or something along
those line).